requirements.txt
aiogram==3.4.0
aiosqlite==0.20.0
aiohttp>=3.9.0,<3.10.0
python-dotenv==1.0.0
